--- a/packages/cms-base/biome.json
+++ b/packages/cms-base/biome.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["../../biome.json"]
+}