--- a/packages/cms-base/components/SwProductCard.vue
+++ b/packages/cms-base/components/SwProductCard.vue
@@ -126,10 +126,10 @@ const toggleWishlistProduct = async () => {
 const addToCartProxy = async () => {
   await addToCart();
   const errors = getErrorsCodes();
-  errors?.forEach((element) => {
+  for (const element of errors) {
     const { messageKey, params } = resolveCartError(element);
     pushError(getCmsTranslate(translations.errors[messageKey], params));
-  });
+  }
 
   if (!errors.length)
     pushSuccess(