--- a/packages/cms-base/components/SwProductListingFilter.vue
+++ b/packages/cms-base/components/SwProductListingFilter.vue
@@ -3,21 +3,22 @@ import SwFilterPropertiesVue from "./listing-filters/SwFilterProperties.vue";
 import SwFilterPriceVue from "./listing-filters/SwFilterPrice.vue";
 import SwFilterRatingVue from "./listing-filters/SwFilterRating.vue";
 import SwFilterShippingFreeVue from "./listing-filters/SwFilterShippingFree.vue";
-
-const emit = defineEmits<{
-  (e: "selectFilterValue", { code, value }: { code: any; value: any }): void;
-}>();
+import type { Component } from "vue";
 
 const props = defineProps<{
   filter: ListingFilter;
   selectedFilters?: {
-    [key: string]: any;
+    [key: string]: unknown;
   };
+}>();
+
+const emit = defineEmits<{
+  selectFilterValue: [{ code: string; value: string | number | boolean }];
 }>();
 
 const cmsMap = () => {
   const map: {
-    [key: string]: any;
+    [key: string]: Component;
   } = {
     manufacturer: SwFilterPropertiesVue,
     properties: SwFilterPropertiesVue,