--- a/packages/cms-base/components/SwSlider.vue
+++ b/packages/cms-base/components/SwSlider.vue
@@ -60,9 +60,7 @@ const children = computed<string[]>(() => {
     ...childrenRaw.value.slice(0, slidesToShow.value),
   ] as string[];
 });
-const emit = defineEmits<{
-  (e: "changeSlide", index: number): void;
-}>();
+const emit = defineEmits<(e: "changeSlide", index: number) => void>();
 const slider = useTemplateRef("slider");
 const imageSlider = useTemplateRef("imageSlider");
 const imageSliderTrackStyle = ref<CSSProperties>();
@@ -115,11 +113,10 @@ const imageSliderStyle = computed(() => {
       height: getConfigValue("minHeight"),
       margin: `0 -${props.gap}`,
     };
-  } else {
-    return {
-      minHeight: getConfigValue("minHeight"),
-    };
   }
+  return {
+    minHeight: getConfigValue("minHeight"),
+  };
 });
 
 const verticalAlignValue = computed(
@@ -147,7 +144,7 @@ function initSlider() {
 
 function buildImageSliderTrackStyle(
   transformIndex: number,
-  moving: boolean = false,
+  moving = false,
   callback = () => {},
 ) {
   let styleObj: CSSProperties = {
@@ -170,8 +167,8 @@ function buildImageSliderTrackStyle(
     imageSliderTrackStyle.value = { ...styleObj };
     isSliding.value = true;
     setTimeout(() => {
-      delete styleObj.transition;
-      imageSliderTrackStyle.value = { ...styleObj };
+      const { transition: _, ...styleWithoutTransition } = styleObj;
+      imageSliderTrackStyle.value = { ...styleWithoutTransition };
       isSliding.value = false;
       callback();
     }, speed.value);
@@ -189,7 +186,7 @@ function buildImageSliderTrackStyle(
       // If image exist
       height = childComponent?.children[0].children[0].clientHeight
         ? `${childComponent.clientHeight}px`
-        : (height = `auto`);
+        : "auto";
     }
     styleObj = {
       ...styleObj,