--- a/packages/cms-base/components/SwVariantConfigurator.vue
+++ b/packages/cms-base/components/SwVariantConfigurator.vue
@@ -31,9 +31,10 @@ let translations: Translations = {
 
 translations = defu(useCmsTranslations(), translations) as Translations;
 
-const emit = defineEmits<{
-  (e: "change", selected: Partial<Schemas["Product"]> | undefined): void;
-}>();
+const emit =
+  defineEmits<
+    (e: "change", selected: Partial<Schemas["Product"]> | undefined) => void
+  >();
 const isLoading = ref<boolean>();
 const router = useRouter();
 const {