--- a/packages/cms-base/components/listing-filters/SwFilterProperties.vue
+++ b/packages/cms-base/components/listing-filters/SwFilterProperties.vue
@@ -19,9 +19,10 @@ const props = defineProps<{
   filter: ListingFilter;
 }>();
 
-const emits = defineEmits<{
-  (e: "select-value", value: { code: string; value: unknown }): void;
-}>();
+const emits =
+  defineEmits<
+    (e: "select-value", value: { code: string; value: unknown }) => void
+  >();
 const selectedOptionIds = inject<string[]>("selectedOptionIds");
 const isFilterVisible = ref<boolean>(false);
 const toggle = () => {