--- a/packages/cms-base/components/listing-filters/SwFilterRating.vue
+++ b/packages/cms-base/components/listing-filters/SwFilterRating.vue
@@ -11,9 +11,10 @@
 import type { Schemas } from "#shopware";
 import { computed, ref } from "vue";
 
-const emits = defineEmits<{
-  (e: "select-value", value: { code: string; value: unknown }): void;
-}>();
+const emits =
+  defineEmits<
+    (e: "select-value", value: { code: string; value: unknown }) => void
+  >();
 
 const props = defineProps<{
   filter: ListingFilter;