--- a/packages/cms-base/components/listing-filters/SwFilterShippingFree.vue
+++ b/packages/cms-base/components/listing-filters/SwFilterShippingFree.vue
@@ -19,9 +19,10 @@ const props = defineProps<{
   selectedFilters: Schemas["ProductListingResult"]["currentFilters"];
 }>();
 
-const emits = defineEmits<{
-  (e: "select-value", value: { code: string; value: unknown }): void;
-}>();
+const emits =
+  defineEmits<
+    (e: "select-value", value: { code: string; value: unknown }) => void
+  >();
 const currentFilterData = computed(
   () => !!props.selectedFilters[props.filter?.code],
 );
@@ -38,7 +39,9 @@ const toggle = () => {
 };
 
 const dropdownElement = ref(null);
-onClickOutside(dropdownElement, () => (isFilterVisible.value = false));
+onClickOutside(dropdownElement, () => {
+  isFilterVisible.value = false;
+});
 </script>
 
 <template>