--- a/packages/cms-base/components/public/cms/CmsGenericElement.vue
+++ b/packages/cms-base/components/public/cms/CmsGenericElement.vue
@@ -17,7 +17,7 @@ const DynamicRender = () => {
   } = resolveCmsComponent(props.content);
   if (resolvedComponent) {
     if (!isResolved)
-      return h("div", {}, "Problem resolving component: " + componentName);
+      return h("div", {}, `Problem resolving component: ${componentName}`);
 
     const { cssClasses, layoutStyles } = getCmsLayoutConfiguration(
       props.content,
@@ -28,7 +28,7 @@ const DynamicRender = () => {
       class: cssClasses,
     });
   }
-  console.error("Component not resolved: " + componentNameToResolve);
+  console.error(`Component not resolved: ${componentNameToResolve}`);
   return h("div", {}, "");
 };
 </script>