--- a/packages/cms-base/components/public/cms/element/CmsElementImageGallery.vue
+++ b/packages/cms-base/components/public/cms/element/CmsElementImageGallery.vue
@@ -77,10 +77,10 @@ function handleChangeSlide(e: number) {
 }
 
 function move(type: "next" | "previous", specificIndex?: number | string) {
-  let step;
+  let step: number;
   const index =
     typeof specificIndex !== "number"
-      ? parseInt(specificIndex as string)
+      ? Number.parseInt(specificIndex as string)
       : specificIndex;
   if (index >= 0) {
     if (type === "next") {