--- a/packages/cms-base/components/public/cms/element/CmsElementProductListing.vue
+++ b/packages/cms-base/components/public/cms/element/CmsElementProductListing.vue
@@ -119,9 +119,7 @@ const compareRouteQueryWithInitialListing = async () => {
   const isChangePageNeeded =
     (route.query.limit && limit.value !== limitListing) ||
     (route.query.p && Number(route.query.p) !== pageListing) ||
-    (route.query.order && route.query.order !== orderListing)
-      ? true
-      : false;
+    (route.query.order && route.query.order !== orderListing);
 
   if (isChangePageNeeded) {
     const limitQuery = route.query.limit