--- a/packages/cms-base/helpers/media/isSpatial.ts
+++ b/packages/cms-base/helpers/media/isSpatial.ts
@@ -4,5 +4,5 @@ export function isSpatial<
     url?: string;
   },
 >(media: T) {
-  return media.fileExtension === "glb" || !!media?.url?.endsWith(".glb");
+  return media.fileExtension === "glb" || !!media.url?.endsWith(".glb");
 }