### Description

switched `cms-base` package linting to biome for better accuracy and performance
test: `time pnpm run lint` inside package

before
```
pnpm run lint  6.63s user 0.69s system 154% cpu 4.725 total
```

after
```
pnpm run lint  0.35s user 0.15s system 85% cpu 0.590 total
```

apart of the speed it was more accurate in finding problems and making the code more consistent