@patzick 2024-12-02
@mkucmus I'll add biome extension to list in VSCode config, but it's not mandatory, more like a helper